advice-yolo/requirements.txt
opencv-python >= 4.5.3
numpy >= 1.19.5
argparse >= 1.1
grpcio >= 1.37.0
grpcio-tools >= 1.37.0
protobuf >= 3.15.8
Pillow